Vorticity Field Animation Frames Generation
I have tried to make this file as generalized as possible but as this is adapted from a program I wrote for my own simulation results (I was simulation the flow of fluid normal to a thin flat plate) there are some elements unique to it. Please read through the code carefully and make the necessary tweaks to ensure it works for your case. I have tried commenting this as best I can but I expect it will not all be easy to understand. So if you have any questions, send me an email at isaac.rosin1@ucalgary.ca.